apps/a2a/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
openai==1.55.3
httpx>=0.27,<1
python-dotenv==1.0.1
pydantic>=2.5.0
# A2A SDK - using HTTP client for now until official SDK is more stable
requests>=2.31.0
websockets>=12.0
